refactor: #980 labels.ts / age-tier.ts の循環依存を解消
UiMode 型・UI_MODES 定数・LEGACY_UI_MODE_MAP・normalizeUiMode を
新設の age-tier-types.ts に集約し、labels.ts と age-tier.ts が共に
このモジュールに依存する形にした。

- labels.ts 内の「意図的に複製」コメントと複製コードを削除
- age-tier.ts は既存 import path 維持のため re-export
- madge --circular で 0 circular を確認済み
- parallel-implementations.md に age-tier-types.ts を追記
