+++ src/lib/domain/labels.ts
@@ -2,35 +2,9 @@
 // 用語辞書 — UI表示ラベルの Single Source of Truth
 // 全てのUIラベルはこのファイルからインポートすること。ハードコード禁止。
 
-import type { UiMode } from './validation/age-tier';
-
-// ============================================================
-// UiMode 正規化（#573: 循環依存を避けるため labels.ts 内に定義）
-// ============================================================
-// NOTE: validation/age-tier.ts の LEGACY_UI_MODE_MAP / normalizeUiMode と
-// 同じ内容を意図的に複製している。age-tier.ts が labels.ts から
-// AGE_TIER_LABELS を import しているため、逆方向の import は循環依存になる。
-// 変更時は両方を同期させること。
-
-const VALID_UI_MODES: ReadonlySet<UiMode> = new Set([
-	'baby',
-	'preschool',
-	'elementary',
-	'junior',
-	'senior',
-]);
-
-const LEGACY_UI_MODE_MAP_LOCAL: Record<string, UiMode> = {
-	kinder: 'preschool',
-	lower: 'elementary',
-	upper: 'junior',
-	teen: 'senior',
-};
-
-function normalizeUiModeForLabel(value: string): UiMode {
-	if (VALID_UI_MODES.has(value as UiMode)) return value as UiMode;
-	return LEGACY_UI_MODE_MAP_LOCAL[value] ?? 'preschool';
-}
+import type { UiMode } from './validation/age-tier-types';
+// #980: age-tier-types.ts に型・正規化関数を集約し循環依存を解消
+import { normalizeUiMode } from './validation/age-tier-types';
 
 // ============================================================
 // ナビゲーションカテゴリ
@@ -98,14 +72,14 @@ export const AGE_TIER_SHORT_LABELS: Record<UiMode, string> = {
  */
 export function getAgeTierLabel(mode: string | null | undefined): string {
 	if (!mode) return AGE_TIER_LABELS.preschool;
-	const normalized = normalizeUiModeForLabel(mode);
+	const normalized = normalizeUiMode(mode);
 	return AGE_TIER_LABELS[normalized];
 }
 
 /** 年齢区分の短縮ラベルを取得（#573: defensive normalization） */
 export function getAgeTierShortLabel(mode: string | null | undefined): string {
 	if (!mode) return AGE_TIER_SHORT_LABELS.preschool;
-	const normalized = normalizeUiModeForLabel(mode);
+	const normalized = normalizeUiMode(mode);
 	return AGE_TIER_SHORT_LABELS[normalized];
 }
